requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pandas==2.1.1
mysql-connector-python==8.0.33
apscheduler==3.10.1
jinja2==3.1.2
requests==2.31.0
thefuzz==0.20.0
faker==19.6.2
